# client/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distill",
    "description": "Distill turns unstructured and semi-structured documents into clean, structured data that can\nbe searched and queried.\n\nThe interesting problems are not extraction from a single document, which modern language\nmodels largely solve, but **unification** (turning many mutually disagreeing documents into\none coherent, queryable schema) and **trust** (making every extracted value traceable to\nwhere it came from, and cheap for a person to correct).\n\nAuthentication is a workspace bearer token, issued once at workspace creation. There are no\naccounts.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}": {
      "patch": {
        "tags": [
          "workspaces"
        ],
        "summary": "Rename a workspace",
        "description": "Set the workspace's name. Requirement FR-01.\n\nThe name is generated once, from the first batch of documents, and is never regenerated\nafterwards \u2014 so this write is final in the sense that matters: nothing else will\noverwrite it.",
        "operationId": "rename_workspace_api_v1_workspaces__workspace_id__patch",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Schema, documents, and counts for a cold start",
        "operationId": "get_workspace_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/records": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Records, cursor paged",
        "operationId": "list_records_api_v1_workspaces__workspace_id__records_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From a previous page.",
              "title": "Cursor"
            },
            "description": "From a previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/records/{record_id}/fields/{field_key}": {
      "patch": {
        "tags": [
          "records"
        ],
        "summary": "Correct one value",
        "description": "Record a human decision about one value. Requirements FR-23 and FR-34.\n\nFrom this point the value is immune to re-extraction: the write path filters\nhuman-owned rows, and a later model answer that disagrees is stored beside it rather\nthan over it (decision D13).",
        "operationId": "correct_field_api_v1_workspaces__workspace_id__records__record_id__fields__field_key__patch",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Record identifier.",
              "title": "Record Id"
            },
            "description": "Record identifier."
          },
          {
            "name": "field_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Key"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldValue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrectionRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "not_present": {
            "type": "boolean",
            "title": "Not Present",
            "description": "The user asserts this document does not contain this field. A correction in its own right, and protected from re-extraction the same way an edited value is.",
            "default": false
          }
        },
        "type": "object",
        "title": "CorrectionRequest",
        "description": "Either a value, or an assertion that the field is absent from this document."
      },
...
      "DocumentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source_format": {
            "$ref": "#/components/schemas/SourceFormat",
            "description": "What the file was actually parsed AS, decided by sniffing the bytes rather than by trusting the extension. The Upload screen shows it per file (decision D44), and showing the extension there instead would hide exactly the case worth seeing: a .csv that is really a tab-separated export, or a .pdf that is a scan.",
            "default": "pdf"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "status",
          "size_bytes",
          "created_at"
        ],
        "title": "DocumentSummary"
      },
      "FieldSpec": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Column name in the generated view. Machine identifier."
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "What the interface shows a person."
          },
          "type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "What this field means. Sent to the model during schema-guided extraction, and used for similarity matching during drift detection.",
            "default": ""
          },
          "enum_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Values",
            "description": "Permitted values, for FieldType.ENUM only."
          },
          "currency_default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency Default",
            "description": "ISO 4217 code assumed when a document states an amount with no currency marker, for FieldType.CURRENCY only."
          },
          "source_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Keys",
            "description": "The raw keys from open extraction that were unified into this field. Kept so the user can see WHY a field exists, and so drift detection knows which aliases have already been resolved."
          },
          "weight": {
            "type": "number",
            "minimum": 0.0,
            "title": "Weight",
            "description": "Review priority multiplier. The review queue orders by weight * (1 - confidence), so a wrong total matters more than a wrong note. Currency and date fields default higher, set in app.schema.propose.",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "type"
        ],
        "title": "FieldSpec",
        "description": "One field in a workspace's schema, as the user has agreed to it."
      },
      "FieldType": {
        "type": "string",
        "enum": [
          "string",
          "number",
          "currency",
          "date",
          "boolean",
          "enum",
          "string_list"
        ],
        "title": "FieldType",
        "description": "The supported field types. Requirement FR-12.\n\nThe stored representation of each is documented in ``app.domain.values``, which is the\nonly module that converts between a raw model output and one of these."
      },
      "FieldValue": {
        "properties": {
          "field_key": {
            "type": "string",
            "title": "Field Key"
          },
          "value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "The value, in the stored representation for its type. Null means the field is genuinely absent from this document, not that extraction failed."
          },
          "value_type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "The model's self-report. Never the tier."
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          },
          "status": {
            "$ref": "#/components/schemas/ValueStatus",
            "default": "model"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Model Value"
          },
          "model_value_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Value At"
          }
        },
        "type": "object",
        "required": [
          "field_key",
          "value_type",
          "tier"
        ],
        "title": "FieldValue",
        "description": "One extracted or corrected value, with everything needed to trust it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Provenance": {
        "properties": {
          "page_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Index",
            "description": "Zero-based page the value was found on, if known."
          },
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/BBox"
            },
            "type": "array",
            "title": "Boxes",
            "description": "One box per visual line of the matched quote. Empty means grounding failed, in which case `failure` says why. Multiple boxes rather than one union box, because a quote that wraps must not highlight the content between its fragments (requirement FR-30)."
          },
          "quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote",
            "description": "The verbatim span the model cited as evidence. Shown to the user in the viewer side card (requirement FR-31)."
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning",
            "description": "The model's short justification for this value. FR-31."
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score",
            "description": "Fuzzy match score of the quote against the page text, 0 to 100. Feeds tier derivation."
          },
          "locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locator",
            "description": "Human-readable address within the source, for formats where one is meaningful: 'Sheet1!B7', 'paragraph 12', 'line 48'. Display only, never parsed."
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundingFailure"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set if and only if `boxes` is empty."
          }
        },
        "type": "object",
        "title": "Provenance",
        "description": "The trail from a value back to the pixels it came from.\n\nAn empty ``boxes`` list with a populated ``failure`` is a legitimate, fully described\nstate: we have a value, we know the model's justification for it, and we are telling the\nuser honestly that we could not find that justification in the document."
      },
      "Record": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name",
            "description": "Original filename. Shown as the row's identity."
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "values": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldValue"
            },
            "type": "object",
            "title": "Values",
            "description": "Keyed by field key. A field absent from this mapping was never extracted for this document, which is different from a field present with a null value, which means the document genuinely does not contain it."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "document_name",
          "schema_version",
          "created_at",
          "updated_at"
        ],
        "title": "Record",
        "description": "One row of the unified table."
      },
      "RecordPage": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/Record"
            },
            "type": "array",
            "title": "Records"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque. Pass back as `cursor` for the next page."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total records in the workspace, for the row count."
          }
        },
        "type": "object",
        "required": [
          "records",
          "total"
        ],
        "title": "RecordPage",
        "description": "A cursor-paged slice of records."
      },
...
      "RenameWorkspaceRequest": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label",
            "description": "What to call this collection. Trimmed; must not be empty, because an empty name is indistinguishable from never having named it, and this route is the only way a person can say what it is called."
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "RenameWorkspaceRequest"
      },
...
      "SourceFormat": {
        "type": "string",
        "enum": [
          "pdf",
          "image",
          "docx",
          "xlsx",
          "csv",
          "text"
        ],
        "title": "SourceFormat",
        "description": "Formats Distill can parse. Requirement FR-02."
      },
...
      "Tier": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low",
          "conflict",
          "verified"
        ],
        "title": "Tier",
        "description": "How much a value should be trusted. Always DERIVED, never asked of the model.\n\nSee decision D5 and the table in implementation.md section 6.4. The derivation lives in\n``app.pipeline.score`` and nowhere else."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValueStatus": {
        "type": "string",
        "enum": [
          "model",
          "human_verified",
          "not_present"
        ],
        "title": "ValueStatus",
        "description": "Where a value came from. Drives the never-lose-a-correction guarantee.\n\nEvery write path a model can reach filters on ``status <> 'human_verified'``. This is\nenforced in SQL rather than by convention, which is what makes principle 4 of\nrequirements.md a guarantee instead of an intention."
      },
      "WorkspaceOverview": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version",
            "description": "Null when no schema has been agreed yet."
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldSpec"
            },
            "type": "array",
            "title": "Fields"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "type": "array",
            "title": "Documents"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count",
            "default": 0
          },
          "dashboard_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashboard Status",
            "description": "pending, ready, or failed. Null when no dashboard row exists yet."
          },
          "dashboard_stale": {
            "type": "boolean",
            "title": "Dashboard Stale",
            "default": false
          },
          "dashboard_panel_count": {
            "type": "integer",
            "title": "Dashboard Panel Count",
            "default": 0
          },
          "last_event_seq": {
            "type": "integer",
            "title": "Last Event Seq",
            "description": "The newest event sequence number. The interface opens its event stream with this as Last-Event-ID, so a cold start streams only what happens NEXT rather than replaying the entire history it has just loaded.",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "created_at"
        ],
        "title": "WorkspaceOverview",
        "description": "Everything the interface needs for a cold start, in one request."
      }
    }
  }
}